add marketcap sum
