+++ src/Dashboard/CoinSummary.js
@@ -1,19 +1,15 @@
-import React from "react";
+import React, {useState} from "react";
 import { Tile } from "../Shared/Tile";
 import { AppContext } from "../App/AppProvider";
-import CoinImage from "../Shared/CoinImage";
 import styled from 'styled-components';
-
-const SpotlightName = styled.h2`
-    text-align: center;
-`
 
 export default function() {
     return (
         <AppContext.Consumer>
-            {({ currentFavorite, coinList }) => (
+            {({ combinedMarketCap, numberFormat }) => (
                 <Tile>
-
+                    <h2>Combined MarketCap</h2>
+                    {numberFormat(combinedMarketCap, 2)}             
                 </Tile>
             )}
         </AppContext.Consumer>
